Move things around 4 / 12
Move ReplAPI implementation to a separate file
